+++ src/components/home/AboutSection.jsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { StaticImage } from 'gatsby-plugin-image';
-import DotsDecoration from './DotsDecoration';
-import DotsRowSmallDecoration from './DotsRowSmallDecoration';
+import DotsDecoration from './decorations/DotsDecoration';
+import DotsRowSmallDecoration from './decorations/DotsRowSmallDecoration';
 
 export default function AboutSection() {
   return (
@@ -11,7 +11,7 @@ export default function AboutSection() {
     >
       <div className='relative '>
         <div className='hidden sm:block sm:absolute sm:w-[95%] sm:h-[95%] sm:top-0 sm:left-0 sm:bg-gradient-to-tr from-standard-blue to-cerulean-crayola z-[-1]' />
-        <div className='sm:absolute sm:w-[95%] sm:h-[95%] sm:bottom-0 sm:right-0 sm:z-10 sm:bg-white sm:p-24 sm:shadow-lg'>
+        <div className='sm:absolute sm:w-[95%] sm:h-[95%] sm:bottom-0 sm:right-0 sm:z-[1] sm:bg-white sm:p-24 sm:shadow-lg'>
           <hgroup className='flex flex-col gap-3 '>
             <h2 className='text-3xl font-bold text-oxford-blue'>
               Nurturing Excellence
