+++ src/components/home/decorations/DotsDecoration.jsx
