+++ src/components/home/decorations/DotsRowSmallDecoration.jsx
