+++ src/components/home/decorations/EventsDecoration.jsx
