+++ src/components/home/decorations/HeroContentDecoration.jsx
